refactor dns out of scanners
import cycling required me to move types/records etc
into their own dnsrecords package
