+++ internal/scanner/scan_dnssec.go
@@ -1,8 +1,9 @@
 package scanner
 
 import (
+	"github.com/danielmichaels/doublestag/internal/dnsclient"
+	"github.com/danielmichaels/doublestag/internal/dnsrecords"
 	"log/slog"
-	"strings"
 
 	"github.com/danielmichaels/doublestag/internal/config"
 	"github.com/danielmichaels/doublestag/internal/logging"
@@ -10,29 +11,15 @@ import (
 )
 
 const (
-	DNSSECNotImplemented           = "DNSSEC not implemented"
-	DNSSECFullyImplemented         = "DNSSEC fully implemented"
-	DNSSECFailedChainOfTrust       = "DNSSEC failed chain of trust validation"
-	DNSSECFailedZoneApexValidation = "DNSSEC failed zone apex validation"
-	DNSSECNotApplicable            = "DNSSEC not applicable"
+	DNSSECNotImplemented   = "DNSSEC not implemented"
+	DNSSECFullyImplemented = "DNSSEC fully implemented"
+	DNSSECNotApplicable    = "DNSSEC not applicable"
 )
-
-type DNSSECResult struct {
-	Domain          string
-	Status          string
-	ValidationError string
-	KSKs            []DNSKEYResult
-	ZSKs            []DNSKEYResult
-	Details         []string
-	HasDNSKEY       bool
-	HasDS           bool
-	HasRRSIG        bool
-}
 
 // todo: a lot of this scanner will need to be moved to an assessor later
 type DNSSECScanner struct {
-	Result *DNSSECResult
-	Client *DNSClient
+	Result *dnsrecords.DNSSECResult
+	Client *dnsclient.DNSClient
 	logger *slog.Logger
 	Domain string
 }
@@ -41,18 +28,18 @@ func NewDNSSECScanner(domain string) *DNSSECScanner {
 	cfg := config.AppConfig()
 	logger, _ := logging.SetupLogger("dns-client", cfg)
 	domain = dns.Fqdn(domain)
-	result := &DNSSECResult{
+	result := &dnsrecords.DNSSECResult{
 		Domain: domain,
 	}
-	client := NewDNSClient()
+	client := dnsclient.NewDNSClient()
 	return &DNSSECScanner{Domain: domain, Client: client, Result: result, logger: logger}
 }
 
-func (s *Scan) ScanDNSSEC(domain string) *DNSSECResult {
+func (s *Scan) ScanDNSSEC(domain string) *dnsrecords.DNSSECResult {
 	ds := NewDNSSECScanner(domain)
 
 	// Check if domain is zone apex
-	if !ds.Client.isZoneApex(domain) {
+	if !ds.Client.IsZoneApex(domain) {
 		ds.Result.Status = DNSSECNotApplicable
 		return ds.Result
 	}
@@ -73,7 +60,7 @@ func (s *Scan) ScanDNSSEC(domain string) *DNSSECResult {
 	dnskeys, rrsigs, dnskeyOK := ds.Client.LookupDNSKEYWithRRSIG(domain)
 	if dnskeyOK {
 		for _, key := range dnskeys {
-			parsedKey, _ := ParseDNSKEY(domain, key)
+			parsedKey, _ := dnsrecords.ParseDNSKEY(domain, key)
 			if parsedKey != nil {
 				if parsedKey.Flags == 257 {
 					ds.Result.KSKs = append(ds.Result.KSKs, *parsedKey)
@@ -96,27 +83,4 @@ func (s *Scan) ScanDNSSEC(domain string) *DNSSECResult {
 	}
 
 	return ds.Result
-}
-
-// compareDS compares two DS records
-func (c *DNSClient) compareDS(ds1, ds2 *dns.DS) bool {
-	// Explicitly lowercase digests for comparison
-	digest1 := strings.ToLower(ds1.Digest)
-	digest2 := strings.ToLower(ds2.Digest)
-
-	result := ds1.KeyTag == ds2.KeyTag &&
-		ds1.Algorithm == ds2.Algorithm &&
-		ds1.DigestType == ds2.DigestType &&
-		digest1 == digest2
-
-	c.logger.Debug("compareDS",
-		"domain", ds1.Hdr.Name,
-		"result", result,
-		"digest1", digest1,
-		"digest2", digest2,
-		"algo1", ds1.Algorithm, "algo2", ds2.Algorithm,
-		"digestType1", ds1.DigestType, "digestType2", ds2.DigestType,
-	)
-
-	return result
 }
